migrate directory page, works
